refactor: merging catalog with posts
